+++ testing/mozharness/scripts/b2g_build.py
@@ -26,7 +26,7 @@ sys.path.insert(1, os.path.dirname(sys.path[0]))
 
 # import the guts
 from mozharness.base.config import parse_config_file
-from mozharness.base.log import WARNING, ERROR, FATAL
+from mozharness.base.log import WARNING, FATAL
 from mozharness.mozilla.l10n.locales import GaiaLocalesMixin, LocalesMixin
 from mozharness.mozilla.purge import PurgeMixin
 from mozharness.mozilla.signing import SigningMixin
@@ -147,7 +147,7 @@ class B2GBuild(LocalesMixin, PurgeMixin,
             'balrog_credentials_file': 'oauth.txt',
             'build_resources_path': '%(abs_obj_dir)s/.mozbuild/build_resources.json',
             'virtualenv_modules': [
-                'requests==2.2.1',
+                'requests==2.8.1',
             ],
             'virtualenv_path': 'venv',
         }
@@ -391,7 +391,7 @@ class B2GBuild(LocalesMixin, PurgeMixin,
             config_dir = os.path.join(dirs['gecko_src'], 'b2g', 'config',
                                       self.config.get('b2g_config_dir', self.config['target']))
             manifest = os.path.abspath(os.path.join(config_dir, gecko_config['tooltool_manifest']))
-            self.tooltool_fetch(manifest=manifest, output_dir=dirs['work_dir'])
+            self.tooltool_fetch(manifest=manifest, output_dir=dirs['gecko_src'])
 
     def unpack_blobs(self):
         dirs = self.query_abs_dirs()
@@ -402,7 +402,7 @@ class B2GBuild(LocalesMixin, PurgeMixin,
             extra_tarballs.extend(gecko_config['additional_source_tarballs'])
 
         for tarball in extra_tarballs:
-            self.run_command(tar + ["xf", tarball], cwd=dirs['work_dir'],
+            self.run_command(tar + ["xf", tarball], cwd=dirs['gecko_src'],
                              halt_on_failure=True, fatal_exit_code=3)
 
     def checkout_gaia_l10n(self):
@@ -785,8 +785,6 @@ class B2GBuild(LocalesMixin, PurgeMixin,
                 self.info("copying %s to public upload directory" % f)
                 self.copy_to_upload_dir(base_f, upload_dir=dirs['abs_public_upload_dir'])
 
-        self.copy_logs_to_upload_dir()
-
     def _do_rsync_upload(self, upload_dir, ssh_key, ssh_user, remote_host,
                          remote_path, remote_symlink_path):
         retval = self.rsync_upload_directory(upload_dir, ssh_key, ssh_user,
@@ -965,11 +963,12 @@ class B2GBuild(LocalesMixin, PurgeMixin,
                     symbols_url = "%s/%s" % (download_url, os.path.basename(matches[0]))
                     downloadables.append(symbols_url)
                     self.set_buildbot_property('symbolsUrl', symbols_url, write_to_file=True)
-                matches = glob.glob(os.path.join(dirs['abs_upload_dir'], 'b2g*tests.zip'))
+                matches = glob.glob(os.path.join(dirs['abs_upload_dir'], 'b2g*test_packages.json'))
                 if matches:
-                    tests_url = "%s/%s" % (download_url, os.path.basename(matches[0]))
-                    downloadables.append(tests_url)
-                    self.set_buildbot_property('testsUrl', tests_url, write_to_file=True)
+                    test_packages_url = "%s/%s" % (download_url, os.path.basename(matches[0]))
+                    downloadables.append(test_packages_url)
+                    self.set_buildbot_property('testPackagesUrl', test_packages_url,
+                                               write_to_file=True)
                     self.invoke_sendchange(downloadables=downloadables)
 
         if self.query_is_nightly() and os.path.exists(dirs['abs_public_upload_dir']) and self.config['upload'].get('public'):
@@ -1147,7 +1146,8 @@ class B2GBuild(LocalesMixin, PurgeMixin,
         self.info("Cleanup .userconfig file.")
         dirs = self.query_abs_dirs()
         userconfig_path = os.path.join(dirs["work_dir"], ".userconfig")
-        os.remove(userconfig_path)
+        if os.path.exists(userconfig_path):
+            os.remove(userconfig_path)
 
 # main {{{1
 if __name__ == '__main__':
